requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.0
numpy==1.14.0
opencv-python==4.2.0.32
pandas==1.0.1
Pillow==7.0.0
pkg-resources==0.0.0
protobuf==3.11.3
Pycasia==0.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
Werkzeug==1.0.0
wrapt==1.12.1
zhon==1.1.5
